Moved my classes to the right folders
