+++ src/main/java/org/example/springboot25/repository/CatPhotoRepository.java
@@ -1,8 +1,7 @@
-package org.example.springboot25.repositories;
+package org.example.springboot25.repository;
 
 import org.example.springboot25.entities.CatPhoto;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 public interface CatPhotoRepository extends JpaRepository<CatPhoto, Long> {
 }
